MouseHandler: route XButtons (Mouse4/5) as keybinds instead of mouse ownership
When bound through uikeys, Mouse4/5 (aka XButtons) try to assign ownership
to gameInputReceiver. This causes either the mouse input or the X-button
bind to be misrouted when pressed in combination.

To prevent this, Mouse4/5 are routed like keybinds instead of sharing mouse
ownership. This allows them to be pressed/released without stealing or
confusing the activeReceiver.